final changes and moving files around
